CI/CD: Fix Clippy errors and integrate quality tools
- Fixed function-to-integer cast errors in `test_suite` and `kernel`.
- Integrated `cargo-llvm-cov` for test coverage reports.
- Added GitHub Pages hosting for coverage reports.
- Added `cargo-audit` for dependency vulnerability scanning.
- Integrated `cargo miri` for the `collections` crate.
- Standardized workspace-wide formatting via `rustfmt`.
- Optimized CI tool installation using `taiki-e/install-action`.
- Configured targeted Clippy suppressions for legacy code.
- Updated .gitignore to exclude local coverage reports.

+++ kernel/src/syscall.rs
@@ -57,12 +57,12 @@ pub fn handle_syscall(num: usize, arg1: usize, arg2: usize, arg3: usize) -> usiz
             crate::keyboard::pop_key().map_or(0, |c| c as usize)
         }
         Ok(SyscallNum::WaitFuture) => {
-            let handle = FutureHandle::unpack(arg1 as usize);
+            let handle = FutureHandle::unpack(arg1);
             let future = kernel().wait_future(handle).unwrap();
             Box::into_raw(Box::new(future)) as usize
         }
         Ok(SyscallNum::IsFutureCompleted) => {
-            let handle = FutureHandle::unpack(arg1 as usize);
+            let handle = FutureHandle::unpack(arg1);
             if kernel().is_future_completed(handle) {
                 1
             } else {
